feat(New Characters) : Add Ghouls, Mortals and Coteries
stylelint linting & standard js fix
