+++ module/actor/coterie-actor-sheet.js
@@ -1,0 +1,244 @@
+/* global DEFAULT_TOKEN, ActorSheet, game, mergeObject, duplicate */
+
+/**
+ * Extend the basic ActorSheet with some very simple modifications
+ * @extends {ActorSheet}
+ */
+
+export class CoterieActorSheet extends ActorSheet {
+  /** @override */
+  static get defaultOptions () {
+    return mergeObject(super.defaultOptions, {
+      classes: ['vtm5e', 'sheet', 'actor', 'coterie'],
+      template: 'systems/vtm5e/templates/actor/coterie-sheet.html',
+      width: 800,
+      height: 700,
+      tabs: [{
+        navSelector: '.sheet-tabs',
+        contentSelector: '.sheet-body',
+        initial: 'features'
+      }]
+    })
+  }
+
+  constructor (actor, options) {
+    super(actor, options)
+    this.locked = true
+    this.isCharacter = false
+  }
+
+  /** @override */
+  get template () {
+    if (!game.user.isGM && this.actor.limited) return 'systems/vtm5e/templates/actor/limited-sheet.html'
+    return 'systems/vtm5e/templates/actor/coterie-sheet.html'
+  }
+
+  /* -------------------------------------------- */
+
+  /** @override */
+  getData () {
+    const data = super.getData()
+    data.locked = this.locked
+    data.isCharacter = this.isCharacter
+    data.sheetType = `${game.i18n.localize('VTM5E.Coterie')}`
+
+    data.dtypes = ['String', 'Number', 'Boolean']
+
+    // Prepare items.
+    if (this.actor.data.type === 'coterie') {
+      this._prepareItems(data)
+    }
+
+    return data
+  }
+
+  /**
+     * Organize and classify Disciplines for Vampire & Ghoul sheets.
+     *
+     * @param {Object} actorData The actor to prepare.
+     *
+     * @return {undefined}
+     */
+  _prepareItems (sheetData) {
+    const actorData = sheetData.actor
+
+    const features = {
+      background: [],
+      merit: [],
+      flaw: []
+    }
+
+    const gear = []
+
+    for (const i of sheetData.items) {
+      i.img = i.img || DEFAULT_TOKEN
+      if (i.type === 'item') {
+        // Append to gear.
+        gear.push(i)
+      } else if (i.type === 'feature') {
+        // Append to features.
+        features[i.data.featuretype].push(i)
+      }
+    }
+
+    actorData.gear = gear
+    actorData.features = features
+  }
+
+  /* -------------------------------------------- */
+
+  /** @override */
+  activateListeners (html) {
+    super.activateListeners(html)
+    this._setupDotCounters(html)
+
+    // Everything below here is only needed if the sheet is editable
+    if (!this.options.editable) return
+
+    // lock button
+    html.find('.lock-btn').click(this._onToggleLocked.bind(this))
+
+    // ressource dots
+    html.find('.resource-value > .resource-value-step').click(this._onDotCounterChange.bind(this))
+    html.find('.resource-value > .resource-value-empty').click(this._onDotCounterEmpty.bind(this))
+
+    // Add Inventory Item
+    html.find('.item-create').click(this._onItemCreate.bind(this))
+
+    // Update Inventory Item
+    html.find('.item-edit').click(ev => {
+      const li = $(ev.currentTarget).parents('.item')
+      const item = this.actor.getOwnedItem(li.data('itemId'))
+      item.sheet.render(true)
+    })
+
+    // Delete Inventory Item
+    html.find('.item-delete').click(ev => {
+      const li = $(ev.currentTarget).parents('.item')
+      this.actor.deleteOwnedItem(li.data('itemId'))
+      li.slideUp(200, () => this.render(false))
+    })
+
+    // Collapsible Features and Powers
+    const coll = document.getElementsByClassName('collapsible')
+    let i
+
+    for (i = 0; i < coll.length; i++) {
+      coll[i].addEventListener('click', function () {
+        this.classList.toggle('active')
+        const content = this.parentElement.nextElementSibling
+        if (content.style.maxHeight) {
+          content.style.maxHeight = null
+        } else {
+          content.style.maxHeight = content.scrollHeight + 'px'
+        }
+      })
+    }
+  }
+
+  _setupDotCounters (html) {
+    html.find('.resource-value').each(function () {
+      const value = Number(this.dataset.value)
+      $(this).find('.resource-value-step').each(function (i) {
+        if (i + 1 <= value) {
+          $(this).addClass('active')
+        }
+      })
+    })
+    html.find('.resource-value-static').each(function () {
+      const value = Number(this.dataset.value)
+      $(this).find('.resource-value-static-step').each(function (i) {
+        if (i + 1 <= value) {
+          $(this).addClass('active')
+        }
+      })
+    })
+  }
+
+  _onToggleLocked (event) {
+    event.preventDefault()
+    this.locked = !this.locked
+    this._render()
+  }
+
+  _onDotCounterChange (event) {
+    event.preventDefault()
+    if (this.locked) return
+    const element = event.currentTarget
+    const dataset = element.dataset
+    const index = Number(dataset.index)
+    const parent = $(element.parentNode)
+    const fieldStrings = parent[0].dataset.name
+    const fields = fieldStrings.split('.')
+    const steps = parent.find('.resource-value-step')
+    if (index < 0 || index > steps.length) {
+      return
+    }
+
+    steps.removeClass('active')
+    steps.each(function (i) {
+      if (i <= index) {
+        $(this).addClass('active')
+      }
+    })
+    this._assignToActorField(fields, index + 1)
+  }
+
+  _onDotCounterEmpty (event) {
+    event.preventDefault()
+    if (this.locked) return
+    const element = event.currentTarget
+    const parent = $(element.parentNode)
+    const fieldStrings = parent[0].dataset.name
+    const fields = fieldStrings.split('.')
+    const steps = parent.find('.resource-value-empty')
+
+    steps.removeClass('active')
+    this._assignToActorField(fields, 0)
+  }
+
+  /**
+   * Handle creating a new Owned Item for the actor using initial data defined in the HTML dataset
+   * @param {Event} event   The originating click event
+   * @protected
+   */
+  _onItemCreate (event) {
+    event.preventDefault()
+    const header = event.currentTarget
+    // Get the type of item to create.
+    const type = header.dataset.type
+    // Grab any data associated with this control.
+    const data = duplicate(header.dataset)
+    // Initialize a default name.
+    const name = `New ${type.capitalize()}`
+    // Prepare the item object.
+    const itemData = {
+      name: name,
+      type: type,
+      data: data
+    }
+    // Remove the type from the dataset since it's in the itemData.type prop.
+    delete itemData.data.type
+
+    // Finally, create the item!
+    return this.actor.createOwnedItem(itemData)
+  }
+
+  // There's gotta be a better way to do this but for the life of me I can't figure it out
+  _assignToActorField (fields, value) {
+    const actorData = duplicate(this.actor)
+    // update actor owned items
+    if (fields.length === 2 && fields[0] === 'items') {
+      for (const i of actorData.items) {
+        if (fields[1] === i._id) {
+          i.data.points = value
+          break
+        }
+      }
+    } else {
+      const lastField = fields.pop()
+      fields.reduce((data, field) => data[field], actorData)[lastField] = value
+    }
+    this.actor.update(actorData)
+  }
+}
