+++ system.json
@@ -2,10 +2,10 @@
   "name": "vtm5e",
   "title": "Vampire the Masquerade 5e",
   "description": "Vampire 5e system for Foundry VTT!",
-  "version": "1.5.1",
+  "version": "1.6.0",
   "minimumCoreVersion": "0.8.0",
   "compatibleCoreVersion": "0.8.6",
-  "templateVersion": 2,
+  "templateVersion": 3,
   "author": "Ray Ji",
   "esmodules": ["module/main.js"],
   "styles": ["css/vtm5e.css"],
