+++ templates/actor/coterie-sheet.html
@@ -1,0 +1,38 @@
+<form class="{{cssClass}} flexcol {{#if locked}}locked{{/if}} coterie-sheet" autocomplete="off">
+
+    {{!-- Coterie Sheet Header --}}
+    <header class="sheet-header">
+        {{> "systems/vtm5e/templates/actor/parts/profile-img.html"}}
+        <div class="header-fields">
+            <div class="grid grid-3col flexrow">
+                <h1 class="charname"><input name="name" type="text" value="{{actor.name}}"
+                    placeholder="{{localize 'VTM5E.Name'}}" /></h1>
+                <h1 class="charname"><input name="data.headers.concept" type="text" value="{{data.data.headers.concept}}"
+                    placeholder="{{localize 'VTM5E.Concept'}}" /></h1>
+                <h1 class="charname"><input name="data.headers.chronicle" type="text" value="{{data.data.headers.chronicle}}"
+                    placeholder="{{localize 'VTM5E.Chronicle'}}" /></h1>
+            </div>
+        </div> 
+    </header>
+
+    {{!-- Coterie Sheet Tab Navigation --}}
+    <nav class="sheet-tabs tabs" data-group="primary">
+        <a class="item" data-tab="features">{{localize "VTM5E.Features"}}</a>
+        <a class="item" data-tab="description">{{localize "VTM5E.Description"}}</a>
+        <a class="item" data-tab="other">{{localize "VTM5E.Other"}}</a>
+    </nav>
+
+    {{!-- Coterie Sheet Body --}}
+    <section class="sheet-body">
+
+        {{!-- Shared Features Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/features.html"}}
+
+        {{!-- Biography Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/biography.html"}}
+
+        {{!-- Other Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/other.html"}}
+        
+    </section>
+</form>
