+++ templates/actor/ghoul-sheet.html
@@ -1,0 +1,61 @@
+<form class="{{cssClass}} flexcol {{#if locked}}locked{{/if}} ghoul-sheet" autocomplete="off">
+
+    {{!-- Ghoul Sheet Header --}}
+    <header class="sheet-header">
+        {{> "systems/vtm5e/templates/actor/parts/profile-img.html"}}
+        <div class="header-fields">
+            <div class="grid grid-3col flexrow">
+                <h1 class="charname"><input name="name" type="text" value="{{actor.name}}"
+                        placeholder="{{localize 'VTM5E.Name'}}" /></h1>
+                <h1 class="charname"><input name="data.headers.domitor" type="text" value="{{data.data.headers.domitor}}"
+                        placeholder="{{localize 'VTM5E.Domitor'}}" /></h1>
+                <div class="resource flex-group-center exp">
+                    {{> "systems/vtm5e/templates/actor/parts/exp.html"}}
+                </div>
+            </div>
+            {{!-- The grid classes are defined in scss/global/_grid.scss. To use,
+            use both the "grid" and "grid-Ncol" class where "N" can be any number
+            from 1 to 12 and will create that number of columns. --}}
+            <div class="resources grid grid-3col">
+                {{!-- "flex-group-center" is also defined in the _grid.scss file
+                and it will add a small amount of padding, a border, and will
+                center all of its child elements content and text. --}}
+
+                {{> "systems/vtm5e/templates/actor/parts/health.html"}}
+                   
+                {{> "systems/vtm5e/templates/actor/parts/willpower.html"}}
+                
+                {{> "systems/vtm5e/templates/actor/parts/humanity.html"}}
+            </div>
+        </div>
+    </header>
+
+    {{!-- Ghoul Sheet Tab Navigation --}}
+    <nav class="sheet-tabs tabs" data-group="primary">
+        <a class="item" data-tab="stats">{{localize "VTM5E.Stats"}}</a>
+        <a class="item" data-tab="disciplines">{{localize "VTM5E.Disciplines"}}</a>
+        <a class="item" data-tab="features">{{localize "VTM5E.Features"}}</a>
+        <a class="item" data-tab="description">{{localize "VTM5E.Description"}}</a>
+        <a class="item" data-tab="other">{{localize "VTM5E.Other"}}</a>
+    </nav>
+
+    {{!-- Ghoul Sheet Body --}}
+    <section class="sheet-body">
+        
+        {{!-- Stats Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/stats.html"}}
+    
+        {{!-- Owned Powers Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/disciplines.html"}}
+
+        {{!-- Owned Features Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/features.html"}}
+
+        {{!-- Biography Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/biography.html"}}
+
+        {{!-- Other Tab --}}
+        {{> "systems/vtm5e/templates/actor/parts/other.html"}}
+        
+    </section>
+</form>
