+++ templates/actor/parts/blood.html
@@ -1,0 +1,43 @@
+<div class="tab blood" data-group="primary" data-tab="blood">
+    <div class="resource flexrow grid grid-2col">
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.BloodPotency"}}: </label>
+            <div class="resource-value" data-value="{{data.data.blood.potency}}" data-name="data.blood.potency">
+                <span class="resource-value-empty"></span>
+                {{#numLoop 10}}
+                <span class="resource-value-step" data-index="{{this}}"></span>
+                {{/numLoop}}
+            </div>
+        </div>
+        <div class="flex-group-center flexrow">
+            <label class="resource-label">{{localize "VTM5E.BloodResonance"}}: </label>
+            <input type="text" name="data.blood.resonance" value="{{data.blood.resonance}}" />
+        </div>
+    </div>
+    <div class="resource flexrow grid grid-2col">
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.BloodSurge"}}:</label>
+            <p>{{lookup blood_potency_text "surge"}}</p>
+        </div>
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.MendAmount"}}:</label>
+            <p>{{lookup blood_potency_text "mend"}}</p>
+        </div>
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.PowerBonus"}}:</label>
+            <p>{{lookup blood_potency_text "power"}}</p>
+        </div>
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.RouseReRoll"}}:</label>
+            <p>{{lookup blood_potency_text "rouse"}}</p>
+        </div>
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.FeedingPenalty"}}:</label>
+            <p>{{lookup blood_potency_text "feeding"}}</p>
+        </div>
+        <div class="flex-group-center">
+            <label class="resource-label">{{localize "VTM5E.BaneSeverity"}}:</label>
+            <p>{{lookup blood_potency_text "bane"}}</p>
+        </div>
+    </div>
+</div>
