+++ templates/actor/parts/exp.html
@@ -1,0 +1,8 @@
+<div class="resource flex-group-center exp">
+    <label for="data.data.exp" class="resource-label">{{localize "VTM5E.Exp"}}</label>
+    <div class="resource-content flexrow flex-center flex-between">
+        <input type="number" name="data.exp.value" value="{{data.data.exp.value}}" data-dtype="Number" />
+        /
+        <input type="number" name="data.exp.max" value="{{data.data.exp.max}}" data-dtype="Number" />
+    </div>
+</div>
