+++ templates/actor/parts/frenzy.html
@@ -1,0 +1,7 @@
+<div class="resource flex-group-center frenzy">
+    <label for="placeholder" class="resource-label">{{localize "VTM5E.Frenzy"}}</label>
+    <div class="resource-content flexrow flex-center flex-between">
+        <button class="rollable" type="text" name="frenzy" data-roll="{{frenzy data.data.willpower.max data.data.willpower.superficial data.data.willpower.aggravated data.data.humanity.value}}"
+            data-label="{{localize 'VTM5E.ResistingFrenzy'}}...">{{localize "VTM5E.Roll"}}</button>
+    </div>
+</div>
