+++ templates/actor/parts/other.html
@@ -1,0 +1,6 @@
+<div class="tab other" data-group="primary" data-tab="other">
+    <label class="resource-label flex-center">{{localize "VTM5E.Notes"}}</label>
+    {{editor content=data.data.notes target="data.notes" button=true owner=owner editable=editable}}
+    <label class="resource-label flex-center">{{localize "VTM5E.Equipment"}}</label>
+    {{editor content=data.data.equipment target="data.equipment" button=true owner=owner editable=editable}}
+</div>
