+++ templates/actor/parts/profile-img.html
@@ -1,0 +1,12 @@
+<div class="profile-img-zone">
+    <h1>{{ sheetType }}</h1>
+    <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" width="100"
+        alt="Profile Image" />
+    <button class="lock-btn">
+        {{#if locked}}
+        <i class="fa fa-lock"></i>
+        {{ else }}
+        <i class="fa fa-lock-open"></i>
+        {{/if}}
+    </button>    
+</div>
