+++ templates/actor/parts/rouse.html
@@ -1,0 +1,7 @@
+<div class="resource flex-group-center rouse">
+    <label for="placeholder" class="resource-label">{{localize "VTM5E.Rouse"}}</label>
+    <div class="resource-content flexrow flex-center flex-between">
+        <button class="rollable" type="text" name="rouse" data-roll="1" data-use-hunger="1"
+            data-label="{{localize 'VTM5E.RousingBlood'}}...">{{localize "VTM5E.Roll"}}</button>
+    </div>
+</div>
