+++ templates/actor/parts/stats.html
@@ -1,0 +1,94 @@
+<div class="tab flex-center" data-group="primary" data-tab="stats">
+    <label class="resource-label">{{localize "VTM5E.Attributes"}}</label>
+    <div class="abilities grid grid-3col flex-group-center">
+        {{#each data.data.abilities as |ability key|}}
+        <div class="ability flexrow flex-center">
+            <label for="data.data.abilities.{{key}}.value" class="resource-label vrollable"
+                data-roll="{{ability.value}}" data-label="{{localize ability.name}}">{{localize
+                ability.name}}</label>
+            <div class="resource-value" data-value="{{ability.value}}" data-name="data.abilities.{{key}}.value">
+                <span class="resource-value-empty"></span>
+                {{#numLoop 5}}
+                <span class="resource-value-step" data-index="{{this}}"></span>
+                {{/numLoop}}
+            </div>
+        </div>
+        {{/each}}
+    </div>
+
+    <label class="resource-label">{{localize "VTM5E.Skills"}}</label>
+    <div class="skills grid grid-3col flex-group-center">
+        {{#each data.data.skills as |skill key|}}
+        <div class="skill flexrow flex-center">
+            <label for="data.data.skills.{{key}}.value" class="resource-label vrollable" data-roll="{{skill.value}}"
+                data-label="{{localize skill.name}}">{{localize skill.name}}</label>
+            <div class="resource-value" data-value="{{skill.value}}" data-name="data.skills.{{key}}.value">
+                <span class="resource-value-empty"></span>
+                {{#numLoop 5}}
+                <span class="resource-value-step" data-index="{{this}}"></span>
+                {{/numLoop}}
+            </div>
+        </div>
+        {{/each}}
+    </div>
+
+    <label class="resource-label">{{localize "VTM5E.Specialties"}}</label>
+    <ol class="items-list flex-left">
+        <li class="item flexrow item-header">
+            <div class="item-image"></div>
+            <div class="item-name">{{localize "VTM5E.Specialty"}}</div>
+            <div class="item-name">{{localize "VTM5E.Skill"}}</div>
+            <div class="item-controls">
+                <a class="item-control item-create" title="{{localize 'VTM5E.AddItem'}}"
+                    data-type="specialty"><i class="fas fa-plus"></i> {{localize "VTM5E.AddItem"}}
+                </a>
+            </div>
+        </li>
+        {{#each actor.specialties as |item id|}}
+        <li class="item flexrow" data-item-id="{{item._id}}">
+            <div class="item-image custom-rollable" data-dice1="" data-dice2="{{item.data.skill}}" data-name="{{item.name}}">
+                <img src="{{item.img}}" title="{{item.name}}" width="24" height="24" alt="{{item.name}}" />
+            </div>
+            <h4 class="item-name">{{item.name}}</h4>
+            <h4 class="item-name">{{localize (generateSkillLabel item.data.skill)}}</h4>
+            <div class="item-controls">
+                <a class="item-control item-edit" title="{{localize 'VTM5E.Edit'}}"><i
+                        class="fas fa-edit"></i></a>
+                <a class="item-control item-delete" title="{{localize 'VTM5E.Delete'}}"><i
+                        class="fas fa-trash"></i></a>
+            </div>
+        </li>
+        {{/each}}
+    </ol>
+
+    <label class="resource-label">{{localize "VTM5E.CustomRolls"}}</label>
+    <ol class="items-list flex-left">
+        <li class="item flexrow item-header">
+            <div class="item-image"></div>
+            <div class="item-name">{{localize "VTM5E.Name"}}</div>
+            <div class="item-name">{{localize "VTM5E.Attribute"}}</div>
+            <div class="item-name">{{localize "VTM5E.Skill"}}</div>
+            <div class="item-controls">
+                <a class="item-control item-create" title="{{localize 'VTM5E.AddItem'}}"
+                    data-type="customRoll"><i class="fas fa-plus"></i> {{localize "VTM5E.AddItem"}}
+                </a>
+            </div>
+        </li>
+        {{#each actor.customRolls as |roll id|}}
+        <li class="item flexrow" data-item-id="{{roll._id}}">
+            <div class="item-image custom-rollable" data-dice1="{{roll.data.dice1}}" data-dice2="{{roll.data.dice2}}" data-name="{{roll.name}}">
+                <img src="{{roll.img}}" title="{{roll.name}}" width="24" height="24" alt="{{roll.name}}" />
+            </div>
+            <h4 class="item-name">{{roll.name}}</h4>
+            <h4 class="item-name">{{localize (generateSkillLabel roll.data.dice1)}}</h4>
+            <h4 class="item-name">{{localize (generateSkillLabel roll.data.dice2)}}</h4>
+            <div class="item-controls">
+                <a class="item-control item-edit" title="{{localize 'VTM5E.Edit'}}"><i
+                        class="fas fa-edit"></i></a>
+                <a class="item-control item-delete" title="{{localize 'VTM5E.Delete'}}"><i
+                        class="fas fa-trash"></i></a>
+            </div>
+        </li>
+        {{/each}}
+    </ol>
+</div>
